pyproject.toml
[project]
name = "fastapi-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.2",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "python-lsp-server>=1.14.0",
    "sqlalchemy>=2.0.48",
]
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
select = ["E", "F", "W", "B", "I", "RUF", "N", "LOG", "ERA", "D", "UP", "ANN",
          "ASYNC", "S", "RET", "TCH", "ARG", "PTH", "DOC"]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
]
